+++ crates/nostr/CHANGELOG.md
@@ -34,6 +34,7 @@
 - Rename `NostrConnectURI` to `NostrConnectUri`
 - Rename `NostrWalletConnectURI` to `NostrWalletConnectUri`
 - Prevent overflow in `Kind` addition
+- Refinements for `wasm32-wasip2` target (https://github.com/rust-nostr/nostr/pull/1164)
 - Make `Kind::from_u16` and `Kind::as_u16` const (https://github.com/rust-nostr/nostr/pull/1136)
 
 ### Added